+++ Controllers/VentasController.cs
@@ -25,9 +25,10 @@ namespace VentaProductos.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Venta>>> GetVenta()
         {
-            var ventas = await _context.Venta.Include(x => x.Cliente).ToListAsync();
+            var ventas = await _context.Venta
+            .Include(x => x.Cliente)
+            .ToListAsync();
             return ventas;
-
         }
 
         // GET: api/Ventas/5
@@ -77,9 +78,20 @@ namespace VentaProductos.Controllers
 
         // POST: api/Ventas
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+
         [HttpPost]
-        public async Task<ActionResult<Venta>> PostVenta(Venta venta)
+        public async Task<ActionResult> PostVenta(Venta venta)
         {
+            var cliente = await _context.Clientes
+                .Include(c => c.Ventas)
+                .FirstOrDefaultAsync(c => c.Id == venta.ClienteId);
+
+            if (cliente != null && cliente.Ventas != null && cliente.Ventas.Any(v => v.Finalizada != true))
+            {
+                // Retornar mensaje de error en el cuerpo de la respuesta
+                return BadRequest(new { message = "El cliente tiene una venta pendiente." });
+            }
+
             _context.Venta.Add(venta);
             await _context.SaveChangesAsync();
 
@@ -95,7 +107,10 @@ namespace VentaProductos.Controllers
             {
                 return NotFound();
             }
-
+            if (!venta.Finalizada)
+            {
+                return BadRequest("No se puede eliminar una venta pendiente.");
+            }
             _context.Venta.Remove(venta);
             await _context.SaveChangesAsync();
